linked allll pages. added footer to app.js, imported header to all pages. folder restructure
