+++ client/src/components/Header.js
@@ -3,9 +3,9 @@ import '../assets/css/style.css';
 
 function Header() {
   return (
-
+    <div style={{padding: 10, backgroundColor: "darkslategray", color: "white"}}>
     <h2>Welcome...(name goes here)</h2>
-    
+    </div>
   );
 }
