+++ client/src/components/pages/Admin.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import '../../assets/css/style.css';
-import StaffCard from '../StaffCard';
-import StudentCard from '../StudentCard';
-import GuardianCard from '../GuardianCard';
-import Footer from "../Footer"
+import StaffCard from '../Cards/StaffCard';
+import StudentCard from '../Cards/StudentCard';
+import GuardianCard from '../Cards/GuardianCard';
+import Header from "../Header"
 import { Link } from "react-router-dom";
 
 
@@ -11,12 +11,12 @@ import { Link } from "react-router-dom";
 function AddUser() {
     return (
     <div>
+        <Header />
         <div className = "container container w3-container w3-center w3-animate-opacity" style={{margin: "auto", width: "75%"}}>
             <StaffCard />
             <StudentCard />
             <GuardianCard />
         </div>
-            <Footer />
     </div>
