+++ client/src/components/pages/History.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import '../../assets/css/style.css';
+import Header from "../Header"
+
+function History() {
+    return (
+        <Header />
+    )
+}
+
+export default History;
