+++ client/src/components/pages/Main.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import '../../assets/css/style.css';
+import Header from "../Header"
+
+function Main() {
+    return (
+        <Header />
+    )
+}
+
+export default Main;
